% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAPMISS.R
\name{FINEMAPMISS}
\alias{FINEMAPMISS}
\title{Fine-mapping meta-analyses with missing information}
\usage{
FINEMAPMISS(
  ses,
  betas,
  R,
  tau = 0.05,
  adj = 1e-04,
  n_reps = 50,
  prob_threshold,
  max_causals = 5,
  cred_sizes = NULL,
  cred_eval,
  meta_analyze = T,
  quiet = T,
  supplied_matrices = NULL,
  adj_match = T,
  RM = NULL,
  estimate_M = FALSE,
  beta_shrink = TRUE,
  n_studies,
  variant_sample_sizes = NULL,
  max_overlap = T,
  freqs,
  scaled_data = F,
  use_N = T
)
}
\arguments{
\item{ses}{A vector or matrix of GWAS marginal effect standard errors.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the standard errors for one dataset. For any unobserved variants in a study,
the standard errors should be set to Inf.}

\item{betas}{A vector or matrix of GWAS marginal effects.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the marginal effects for one dataset. For any unobserved variants in a study,
the marginal effects should be set to 0.}

\item{R}{Reference LD matrix for the set of analyzed variants.}

\item{tau}{Prior standard error for the casual effects. Set by default to 0.05.}

\item{adj}{Adjustment constant \eqn{\varepsilon}, shrinks the correlations in
\eqn{\boldsymbol{R_M}} to make its inversion stable.}

\item{n_reps}{Maximum number of Stochastic Shotgun Search (SSS) iterations.}

\item{prob_threshold}{SSS termination threshold, based on probability mass added at each iteration.}

\item{max_causals}{Maximum number of causal variants.}

\item{cred_sizes}{Vector or numeric of size of credible sets to be evaluated. If no value
is provided, the rounded expected posterior for number of causal variants is used.}

\item{cred_eval}{Binary variable, should credible sets be computed?}

\item{meta_analyze}{Binary variable, should the data be meta-analyzed? If data has been
combined in advance, should be set to \code{FALSE}.}

\item{quiet}{Should the current progress of SSS be repressed?}

\item{supplied_matrices}{If the vector \eqn{\boldsymbol{\widehat z}^{T} \boldsymbol{R_M}^{-1}\boldsymbol{R}_{\boldsymbol{\gamma}} }
and matrix \eqn{\mathbb{I}_p + \boldsymbol{R}_{\boldsymbol{\gamma}}^{T} \boldsymbol{R_M}^{-1} \boldsymbol{R}_{\boldsymbol{\gamma}} }
have been computed in advance, they can be input as a list to avoid computing them again.}

\item{adj_match}{Should the \eqn{\boldsymbol{R}} be adjusted in a similar way to \eqn{\boldsymbol{R_M}} when computing log Bayes-factors?}

\item{RM}{If \eqn{\boldsymbol{R_M}} has been computed in advance, it can be provided as input.}

\item{estimate_M}{Should \eqn{\boldsymbol{M}} be estimated? This should be used if data has been meta-analyzed in advance.}

\item{beta_shrink}{Should beta shrinkage in the
marginal effect covariances be applied?}

\item{n_studies}{Number of studies. Should be equal to the number of columns in
ses abnd betas. If the data has been combined in advance, set to 1.}

\item{variant_sample_sizes}{Vector of combined sample sizes per variant, from the meta-analysis.}

\item{max_overlap}{Binary parameter denoting whether the maximum sample
overlap is assumed when estimating \eqn{\boldsymbol{M}}.}

\item{freqs}{A vector or matrix of variant frequencies.
Takes the form of a vector if there is only a single dataset or the data has
been combined in advance. Otherwise, it is a matrix where each column contains
the variant frequencies for one dataset. For any unobserved variants in a study,
the marginal effects should be set to 0.}

\item{scaled_data}{Has the data been scaled with allele frequencies in advance?}

\item{use_N}{Are the variant sample sizes or marginal effect standard errors used
for fine-mapping?}
}
\value{
A list of objects
\describe{
  \item{\code{pips}}{Numeric vector of posterior inclusion probabilities per variant.}
  \item{\code{stored_configs}}{A matrix of evaluated configurations and associated information.
   Column 1. causal configuration, 2. log Bayes-factor, 3. size of configuration, 4. Is this a unique configuration?}
  \item{\code{cred}}{List or matrix of credible sets.}
  \item{\code{post_k}}{Posterior probability distribution for the number of causal variants from \code{0:max_causals}}
}
}
\description{
\code{FINEMAPMISS} performs Bayesian variable selection on a set of genetic
variants and a phenotype, using marginal effect estimates \eqn{\boldsymbol{\widehat \beta}},
 their standard errors \eqn{\boldsymbol{s}}, and  a reference LD panel \eqn{\boldsymbol{R}}
 as input. \cr
 \code{FINEMAPMISS} differs from previous methods, as it allows for the combination
 of multiple datasets across biobanks, while also functioning with missing data. \cr
 If given separate data, this function will automatically perform the meta-analysis
 and fine-map the data. \cr
 If the data have been combined into a single dataset in advance, \code{FINEMAPMISS}
 can still be run, but with reduced accuracy.
}
