% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_summary_statistics.R
\name{impute_summary_stats}
\alias{impute_summary_stats}
\title{Impute summary statistics}
\usage{
impute_summary_stats(
  R,
  z,
  typed,
  untyped,
  shrink = 1e-04,
  return_z = TRUE,
  scale = T,
  n = NULL
)
}
\arguments{
\item{R}{LD matrix}

\item{z}{z-scores}

\item{typed}{Index vector of observed variants}

\item{untyped}{Index vector of unobserved variants}

\item{shrink}{Shrinkage factor}

\item{return_z}{Return results as z-scores?}

\item{scale}{Scale z-scores after imputation?}

\item{n}{Sample size}
}
\value{
Matrix with columns: 1. Imputed z-scores (marginal effect if return_z == F), 2. Imputation quality, 3. (if return_z == F) standard error.
}
\description{
Impute summary statistics
}
